factor the slices stuff into slice.rs
The lib.rs is somewhat unwieldly, putting things into dedicate files makes easier to
handle (IMO).

Eventually this should be done with the iterator into a 'iter.rs' as well.
